Some 64-bit compilation fixes (hopefully this didn't break everything else)

+++ EncryptedStream.h
@@ -15,13 +15,13 @@ public:
 	EncryptedStream(Stream *stream);
 	virtual ~EncryptedStream();
 
-	virtual ssize_t Read(void* dst, int size);
-	virtual ssize_t ReadAt(int pos, void *dst, int size);
+	virtual ssize_t Read(void* dst, size_t size);
+	virtual ssize_t ReadAt(off_t pos, void *dst, size_t size);
 
-	virtual int32 Seek(int32 where, int whence);
-	virtual int32 Position() const;
+	virtual off_t Seek(off_t where, int whence);
+	virtual off_t Position() const;
 
-	virtual uint32 Size() const;
+	virtual size_t Size() const;
 
 private:
 	Stream* fEncryptedStream;
