+++ streams/Stream.h
@@ -10,29 +10,29 @@ public:
 	Stream();
 	virtual ~Stream();
 	
-	virtual ssize_t Read(void *dst, int size);
-	virtual ssize_t ReadAt(int pos, void *dst, int size) = 0;
+	virtual ssize_t Read(void *dst, size_t size);
+	virtual ssize_t ReadAt(off_t pos, void *dst, size_t size) = 0;
 
 	char *ReadLine(char *buffer, size_t maxSize, char endLine = '\n');
 	ssize_t ReadString(char *string, size_t size);
 
-	virtual ssize_t Write(const void *src, int size);
-	virtual ssize_t WriteAt(int pos, const void *src, int size)
+	virtual ssize_t Write(const void *src, size_t size);
+	virtual ssize_t WriteAt(off_t pos, const void *src, size_t size)
 	{
 		throw "Stream is not writable!";
 	};
 	
-	virtual int32 Seek(int32 where, int whence)
+	virtual off_t Seek(off_t where, int whence)
 	{
 		throw "Seek() not supported for this stream!";
 	};
 	
-	virtual int32 Position() const 
+	virtual off_t Position() const 
 	{
 		throw "Position() not supported for this stream!";
 	};
 	
-	virtual uint32 Size() const
+	virtual size_t Size() const
 	{
 		throw "Size() not supported for this stream!";
 	}
@@ -60,7 +60,7 @@ public:
 	}
 	
 	template<typename T>
-	void ReadAt(int pos, T &dst) {
+	void ReadAt(off_t pos, T &dst) {
 		ssize_t read = ReadAt(pos, &dst, sizeof(dst));
 		if (read < 0 || read != sizeof(dst))
 			throw "ReadAt() exception";
